add authentification

+++ src/OC/PlatformBundle/Controller/AdvertController.php
@@ -10,7 +10,9 @@ use OC\PlatformBundle\Entity\Application;
 use OC\PlatformBundle\Entity\Image;
 use OC\PlatformBundle\Form\AdvertEditType;
 use OC\PlatformBundle\Form\AdvertType;
+use Sensio\Bundle\FrameworkExtraBundle\Configuration\Security;
 use Symfony\Bundle\FrameworkBundle\Controller\Controller;
+use Symfony\Component\Finder\Exception\AccessDeniedException;
 use Symfony\Component\Form\Extension\Core\Type\CheckboxType;
 use Symfony\Component\Form\Extension\Core\Type\DateType;
 use Symfony\Component\Form\Extension\Core\Type\FormType;
@@ -28,15 +30,12 @@ class AdvertController extends Controller
         // On ne sait pas combien de pages il y a
         // Mais on sait qu'une page doit être supérieure ou égale à 1
 
-
         if ($page < 1) {
             // On déclenche une exception NotFoundHttpException, cela va afficher
             // une page d'erreur 404 (qu'on pourra personnaliser plus tard d'ailleurs)
             throw new NotFoundHttpException('Page "'.$page.'" inexistante.');
         }
-
         // Ici, on récupérera la liste des annonces, puis on la passera au template
-
         // Mais pour l'instant, on ne fait qu'appeler le template
 
         $em = $this->getDoctrine()->getManager();
@@ -100,11 +99,21 @@ class AdvertController extends Controller
     }
 
 
+    /**
+     * @Security("has_role('ROLE_AUTEUR')")
+     * @param Request $request
+     * @return \Symfony\Component\HttpFoundation\RedirectResponse|\Symfony\Component\HttpFoundation\Response
+     */
     public function addAction(Request $request)
     {
         // La gestion d'un formulaire est particulière, mais l'idée est la suivante :
 
-        // Si la requête est en POST, c'est que le visiteur a soumis le formulaire
+        // Si la requête est en POST, c'est que le visiteur a soumis le formulair
+
+        if(!$this->get('security.authorization_checker')->isGranted('ROLE_AUTEUR')){
+            throw new AccessDeniedException('acces limite aux auteurs');
+        }
+
         $advert = new Advert();
         $form= $this->get('form.factory')->create(AdvertType::class, $advert);
 
@@ -130,8 +139,17 @@ class AdvertController extends Controller
         ));
     }
 
+    /**
+     * @Security("has_role('ROLE_MODERATEUR')")
+     * @param $id
+     * @param Request $request
+     * @return \Symfony\Component\HttpFoundation\RedirectResponse|\Symfony\Component\HttpFoundation\Response
+     *
+     */
     public function editAction($id, Request $request)
     {
+
+
         $em = $this->getDoctrine()->getManager();
 
         $advert = $em->getRepository("OCPlatformBundle:Advert")
@@ -154,7 +172,7 @@ class AdvertController extends Controller
         ));
     }
 
-    public function deleteAction($id)
+    public function deleteAction(Request $request, $id)
     {
         $em = $this->getDoctrine()->getManager();
 
@@ -165,13 +183,25 @@ class AdvertController extends Controller
             throw new NotFoundHttpException("not advert found");
         }
 
-        foreach($advert->getCategories() as $value){
-            $advert->removeCategory($value);
-        }
-        $em->remove($advert);
-        $em->flush();
+        $form = $this->get('form.factory')->create();
 
-        return $this->render('OCPlatformBundle:Advert:delete.html.twig');
+        if ($request->isMethod('POST') && $form->handleRequest($request)->isValid()) {
+            foreach($advert->getCategories() as $value){
+                $advert->removeCategory($value);
+            }
+            $em->remove($advert);
+            $em->flush();
+
+            $request->getSession()->getFlashBag()->add('info', "L'annonce a bien été supprimée.");
+
+            return $this->redirectToRoute('oc_platform_home');
+        }
+
+
+
+        return $this->render('OCPlatformBundle:Advert:delete.html.twig',
+            array('advert'=>$advert,
+                  'form'  =>$form->createView()));
     }
 
 }
